requirements.txt
Flask==2.3.2
flask-cors==4.0.0
tensorflow==2.13.0
Pillow==10.0.0
gunicorn==21.2.0
